feat(api): removed unnecessary object wrapping
refs. metriport/metriport#426

+++ api/app/src/command/connected-user/delete-connected-user.ts
@@ -1,5 +1,5 @@
 import { ConnectedUser } from "../../models/connected-user";
 
-export const deleteConnectedUser = async ({ id }: { id: string }): Promise<void> => {
-  await ConnectedUser.destroy({ where: { id } });
+export const deleteConnectedUser = async (userId: string): Promise<void> => {
+  await ConnectedUser.destroy({ where: { id: userId } });
 };
